Move realmd packets to file `AuthPackets.h`
